Move letter behind not android target flag
