+++ server/service/Cargo.toml
@@ -61,6 +61,7 @@ rust-embed = { version = "8.4.0", features = ["include-exclude"] }
 extism = { workspace = true }
 base64 = "0.22.1"
 regex = { workspace = true }
+[target.'cfg(not(target_os = "android"))'.dependencies]
 lettre = "0.11.11"
